(SLV-261) move to bolt 1.0
This is not an externally releaseable version

This required switching to using bolt as a package because bolt 1.0 installs a puppet gem on the path, which the installer happily used... and failed to install.
Since the bolt package has a ruby version, we should use that to for RAS rather than make the user install ruby.
That ran into the problem that RAS depends on oga which has native extensions that need to be built.  And gcc may not be on the machine.
So I built a version of the bolt package with oga.  Put that in fixtures to use for acceptance tests.
That is why this isn't releasable.  We need to either get Bolt to include oga official (and it's deps) or maybe even move all of ras into bolt.  But that will come later.

Also, the bolt package had the facts module we were downloading before.  So removed code for downloading that.  Though the version there is a little different, the plan we were using is gone, and we had to call a task directly.  The task output is a little different unless we add --format json to the task runs.  So added that to all task runs as really if we are going to parse the output we will want it in json anyway.
That caused update to unit tests.

Also to do much of the above, had to change several of the pre-suites.

Oh, and the PE uninstaller has a "bug" where it breaks bolt, so our teardowns have to reinstall bolt and ras if running on the controller.

oh and the bolt.yaml location changed with 1.0...

yeah this was fun peeling the onion.